relax /system/bin directory permissions
In commit 9e9aa6964a814fc2a7c00d5f4dc4522bac0d8092 [formerly f4fc922f0b863659ca8e97c1f5fa522fafc7deb6], we tightened the
permissions on various bin directories. Please see
https://android-review.googlesource.com/c/platform/system/core/+/822955
for details.

This change causes the Chase banking app to crash. This is because
the Chase app is using inotify_add_watch() on the /system/bin directory
and not checking the return value.

The Android Security model guarantees the immutability of files in
/system/bin, so the inotify watch is unnecessary.

Until the Chase app fixes their bug, we need to relax the permissions on
the /system/bin directory. Conceptually, this is a partial revert of
9e9aa6964a814fc2a7c00d5f4dc4522bac0d8092 [formerly f4fc922f0b863659ca8e97c1f5fa522fafc7deb6].

Bug: 119605322
Test: compiles
Change-Id: Ic72dd24cb27cff677093963bdfd0ae09bf132e08